Tests/LibWeb/Text/expected/css/stylesheet-remove-pseudo-element-only-invalidation.txt

before color: rgb(255, 0, 0)
before background: rgb(0, 255, 0)
after color: rgb(0, 0, 0)
after background: rgba(0, 0, 0, 0)
